+++ NEWS.md
@@ -1,3 +1,9 @@
+# spant 2.23.0
+* abfit options for lipid and MM maximum broadening and maximum frequency shift
+now default to be the same as the metabolite values unless explicitly changed.
+* Added function to calculate the correlation matrix from a basis_set.
+* Fix for RDA files where a comma is sometimes used as a decimal point.
+
 # spant 2.22.0
 * Added find_bids_mrs and mrs_data2bids helper functions.
 * Improved handling of CMRR semi-LASER SVS water reference scans.